requirements.txt
Django==1.10.1
djangorestframework==3.5.3
Pillow==3.4.2
